Other Continuous Random Variables

Exponential

RandomVariates.expon_rng — Function

expon_rng(λ, size=1; seed=nothing)

Generate a size element array of random variables from a Exponential(λ) distribution. Optionally you can set a specific seed.

Examples

julia> expon_rng(3)
1-element Vector{Float64}:
 0.07033135663980515

julia> expon_rng(1.2, seed=42)
1-element Vector{Float64}:
 0.3296112244200808

julia> expon_rng(1.2, (2, 2))
2×2 Matrix{Float64}:
 1.9327    0.134739
 0.746861  0.155614

Erlang

RandomVariates.erlang_rng — Function

erlang_rng(k, λ, size=1; seed=nothing)

Generate a size element array of random variables from a Erlang_{k}(λ) distribution. Optionally you can set a specific seed.

Examples

julia> erlang_rng(5, .5)
1-element Vector{Float64}:
 10.803989701023117

julia> erlang_rng(3, 1, (2,2))
2×2×1 Array{Float64, 3}:
[:, :, 1] =
 2.19956  4.18505
 5.46892  2.5633

Weibull

RandomVariates.weibull_rng — Function

weibull_rng(λ, β, size=1; seed=nothing)

Generate a size element array of random variables from a Weibull(λ, β) distribution. Optionally you can set a specific seed.

Examples

julia> weibull_rng(2, 2, seed=42)
1-element Vector{Float64}:
 0.31445725834527055

julia> weibull_rng(2, 2, 2)
2-element Vector{Float64}:
 0.39561285703154575
 0.6021921673483441

julia> weibull_rng(2, 2, (2,2))
2×2 Matrix{Float64}:
 0.428896  0.109897
 0.812854  0.427906
 

Gamma

RandomVariates.gamma_rng — Function

gamma_rng(α, β, size=1; seed=nothing)

Generate a size element array of random variables from a Gamma(α, β) distribution. Optionally you can set a specific seed.

Examples

julia> gamma_rng(1,1)
1-element Vector{Float64}:
 0.5190236735858542

julia> gamma_rng(1,1,4)
4-element Vector{Float64}:
 0.3035517926878862
 0.5765419737109622
 0.44121996206333797
 0.7325887616559309

julia> gamma_rng(1,1,(2,2))
2×2 Matrix{Float64}:
 0.228818  0.88849
 0.665729  1.01668
 

Beta

RandomVariates.beta_rng — Function

beta_rng(α, β, size=1; seed=nothing)

Generate a size element array of random variables from a Beta(α, β) distribution. Optionally you can set a specific seed.

Examples

julia> beta_rng(1,2)
1-element Vector{Float64}:
 0.44456674672905633

julia> neg_binomial_rng(.5, 5, 5)
5-element Vector{Float64}:
  8.0
 10.0
  8.0
 13.0
 10.0

julia> neg_binomial_rng(.5, 2, (2,2))
2×2 Matrix{Float64}:
 3.0  4.0
 4.0  2.0