Normal Random Variables

Generating Standard Normal

RandomVariates.get_std_normal — Function

get_std_normal(size=1; seed=nothing)

Generate a size element array of random variables from a standard Normal(0, 1) distribution. Optionally you can set a specific seed.

Examples

julia> get_std_normal()
1-element Vector{Float64}:
 0.6315076033452351

julia> get_std_normal(5, seed=43)
5-element Vector{Float64}:
  1.2311463458421277
  1.7786409025309897
 -0.4178415161339713
  0.3518755172644067
 -0.16742990320047046

julia> get_std_normal((2,2))
2×2 Matrix{Float64}:
 -0.900365   -0.432759
 -0.0350299   1.55754

Normal

RandomVariates.normal_rng — Function

normal_rng(μ, σ², size=1; seed=nothing)

Generate a size element array of random variables from a Normal(μ, σ²) distribution. Optionally you can set a specific seed.

Notes

The Normal distribution is given by:

$f(x, μ, σ²) = \frac{1}{σ\sqrt{2π}} e^{ - \frac{1}{2} (\frac{x-μ}{σ})^2}$

Examples

julia> normal_rng()
1-element Vector{Float64}:
 0.03130435813519526

julia> normal_rng(3, 9, 2)
2-element Vector{Float64}:
  7.362935421449054
 -1.0173543995738399

julia> normal_rng(0,1,(2,2))
2×2 Matrix{Float64}:
 -0.640505   0.30303
 -0.0556832  0.714122
 

References

Walk, C. Handbook on statistical distributions for experimentalists. 2007.